requirements.txt
flask
python-dotenv
python-telegram-bot==13.14
openai>=1.0.0
requests
beautifulsoup4
urllib3==1.26.16
